[PAM-925] flyttet util-funksjon
