+++ src/search/facets/sector/sectorReducer.js
@@ -1,5 +1,5 @@
 import { SET_INITIAL_STATE, FETCH_INITIAL_FACETS_SUCCESS, SEARCH_SUCCESS } from '../../searchReducer';
-import { moveFacetToBottom} from "../../../utils";
+import { moveFacetToBottom} from "../utils";
 
 export const CHECK_SECTOR = 'CHECK_SECTOR';
 export const UNCHECK_SECTOR = 'UNCHECK_SECTOR';
