+++ src/utils.js
@@ -28,17 +28,4 @@ export function formatISOString(isoString, format = 'MMMM YYYY') {
     return '';
 }
 
-export function moveFacetToBottom(array, itemKey){
-    const clone = array.concat();
-
-    clone.forEach(function (item, i) {
-        if(item.key === itemKey) {
-            // Flytter elementet til bunn av listen
-            clone.push(clone.splice(i, 1)[0]);
-        }
-    });
-
-    return clone;
-};
-
 export default toDate;
